+++ src/membuf.c
@@ -166,7 +166,7 @@ xmlSecTransformMemBufExecute(xmlSecTransformPtr transform, int last, xmlSecTrans
         if(ret < 0) {
             xmlSecInternalError2("xmlSecBufferAppend",
                                  xmlSecTransformGetName(transform),
-                                 "size=%d", inSize);
+                                 "size=%d", (int)inSize);
             return(-1);
         }
 
@@ -174,7 +174,7 @@ xmlSecTransformMemBufExecute(xmlSecTransformPtr transform, int last, xmlSecTrans
         if(ret < 0) {
             xmlSecInternalError2("xmlSecBufferAppend",
                                  xmlSecTransformGetName(transform),
-                                 "size=%d", inSize);
+                                 "size=%d", (int)inSize);
             return(-1);
         }
 
@@ -182,7 +182,7 @@ xmlSecTransformMemBufExecute(xmlSecTransformPtr transform, int last, xmlSecTrans
         if(ret < 0) {
             xmlSecInternalError2("xmlSecBufferRemoveHead",
                                  xmlSecTransformGetName(transform),
-                                 "size=%d", inSize);
+                                 "size=%d", (int)inSize);
             return(-1);
         }
