+++ src/relationship.c
@@ -689,7 +689,7 @@ xmlSecTransformRelationshipPopBin(xmlSecTransformPtr transform, xmlSecByte* data
            if(ret < 0) {
                xmlSecInternalError2("xmlSecBufferRemoveHead",
                                     xmlSecTransformGetName(transform),
-                                    "size=%d", outSize);
+                                    "size=%d", (int)outSize);
                return(-1);
            }
        } else if(xmlSecBufferGetSize(out) == 0) {
